updated typing

+++ jenkins/pipelines/packer/packer-build.groovy
@@ -143,19 +143,25 @@ pipeline {
         stage("Cleanup Old Images") {
             steps {
                 script {
-                    String output = sh(returnStdout: true, script: "ls -lt /opt/vm-images/")
+                    String output = sh(returnStdout: true, script: "ls -lt /opt/vm-images/${codeUnit.name}*")
                     List<String> imageNames = []
                     String[] lines = output.split('\n')
-                    for (int i = 0; i < Math.min(lines.length - 8, lines.length); i++) {
+                    for (int i = 0; i < Math.min(lines.length - 8, lines.length - 1); i++) {
                         String line = lines[i]
-                        String[] lineParts = line.split(' ')
-                        String imagePath = lineParts[8]
-                        String[] imagePathParts = imagePath.split('/')
-                        imageName = imagePathParts[imagePathParts.length - 1]
+                        if (line.startsWith("total")) {
+                            continue
+                        }
+                        String imagePath = line.split(' ')
+                                .find({ part -> part.contains("qcow2") || part.contains("sha512") })
+                        imageNames.add(
+                                imagePath.split('/').find({ part -> part.contains("qcow2") || part.contains("sha512") })
+                        )
                     }
                     imageNames.each { toBeRemoved ->
-                        print("Removing /opt/vm-images/${toBeRemoved}")
-                        sh "rm /opt/vm-images/${toBeRemoved}"
+                        if (StringUtils.isNotBlank(toBeRemoved.trim())) {
+                            print("Removing /opt/vm-images/${toBeRemoved}")
+                            sh "rm /opt/vm-images/${toBeRemoved}"
+                        }
                     }
                 }
             }
